moved logout window to dialog folder and moved css
