+++ Antrashare_ui/src/app/css/logout-window.component.scss
