+++ Antrashare_ui/src/app/dialogs/logout-window/logout-window.component.ts
@@ -4,7 +4,7 @@ import { Router } from '@angular/router';
 @Component({
   selector: 'app-logout-window',
   templateUrl: './logout-window.component.html',
-  styleUrls: ['./logout-window.component.scss']
+  styleUrls: ['../../css/logout-window.component.scss']
 })
 export class LogoutWindowComponent implements OnInit {
